fix(DPOOC2-104): Fix problem with ShareMapper
There will be a separate ShareMapper for Client and another one for Server.
The reason why you needed to include the dependencies server-client in the
first place is because of this. But as you can see, it gave errors.
Therefore, by following the pattern, we have to create a separate on for
the Client, which is exactly the same because our model entities of Share and
Company are the same for both Server and Client. In real-life projects, this
doesn't always happen. Therefore, the concept of mappers and their
implementation is most commonly used.

Moreover, to be able to use the mapper, we don't create static methods in the
interface. We create an instance of the interface implementation
(ShareMapperImpl) and add it as an attribute to NetworkManager to be able to
use the methods in the implementation.